Pick up biasadj from existing model
Currently if an existing model has the bias adjustment parameter set, then this is dropped when passing the model to Arima. This change checks if the lambda already has a bias adjustment set (since lambda parameter defaults to coming from the original model). Not sure if this change should be implemented in arima2 instead.