Make allseas in mstl.R a ts object
This prevents an error that can occur when there are missing values at the
beginning of the response variable and the auto.arima model function is used.
In that case, the fitted time series starts at the first non-missing value.
Consequently, it is not of the same length as allseas, and adding them together
when allseas is not a ts object results in an error.