Subset xreg along with x
The documentation says that xreg should have the same length or number of rows as x.
But when x beings with NAs, the leading NAs get trimmed. This leads to errors unless
xreg is actually trimmed by the user. To prevent this extra work for the user, we
subset x and xreg together when there are leading NAs in x.